backend/pyproject.toml
[project]
name = "postiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "psycopg[binary]>=3.2.9",
    "python-jose>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
    "passlib[argon2]>=1.7.4",
    "python-multipart>=0.0.20",
    "openai>=2.30.0",
]
